Initial Commit - Gamemode Release
This is more intended for developers to build on but you could totally play with some friends with this. its kinda basic. you can choose teams. if you fall of the map you die. theres also no vehicles enabled here and weapons are chosen by team. its all client side as well so there may be bugs i havent found in MP situations. 

if you find a bug report a problem here and ill fix it or help you fix it | merge commits that fix it lmao |